Added new burn no upsell test
